Resolve "Clean up mongoose interface typing"

Closes #31

See merge request segfault/segfault-backend!18

+++ src/services/board.ts
@@ -4,7 +4,7 @@ import BoardCategoryModel from "../models/boardCategory";
 import HttpError from "../utils/httpError";
 
 /**
- * Get all boards from the database
+ * Fetch a list containing all boards in the database
  *
  * @returns An array of board objects
  */
@@ -21,9 +21,9 @@ const getAll = async () => {
 };
 
 /**
- * Get a board from the database by its id
+ * Fetch a board from the database based on its id
  *
- * @param id The id of the board to fetch
+ * @param {string} id - The id of the board to fetch
  *
  * @returns An object containing board information
  */
@@ -33,10 +33,8 @@ const getById = async (id: string) => {
   try {
     board = await BoardModel.findById(id);
   } catch (err) {
-    console.log(err);
     throw new HttpError("Error searching for board", 500);
   }
-
   if (!board) {
     throw new HttpError("Board not found", 404);
   }
@@ -47,10 +45,10 @@ const getById = async (id: string) => {
 /**
  * Update a board's information
  *
- * @param id The id of the board to update
- * @param topic The new topic for the board
- * @param description The new description for the board
- * @param categoryId The id of the category to assign the board to
+ * @param {string} id - The id of the board to update
+ * @param {string} topic - The new topic for the board
+ * @param {string} description - The new description for the board
+ * @param {string} categoryId - The id of the category to assign the board to
  *
  * @returns An object containing updated board information
  */
@@ -66,7 +64,6 @@ const update = async (
   } catch (err) {
     throw new HttpError("Error fetching board", 500);
   }
-
   if (!board) {
     throw new HttpError("Board not found", 404);
   }
@@ -78,23 +75,34 @@ const update = async (
     const sess = await mongoose.startSession();
     sess.startTransaction();
 
-    if (categoryId && categoryId !== board.category.id.toString()) {
+    if (categoryId && categoryId !== board.category.categoryId.toString()) {
       // Remove from original category
       const boardId = board.id;
-      const oldCategory = await BoardCategoryModel.findById(board.category.id);
-      const idx = oldCategory?.boards.findIndex(
+      const oldCategory = await BoardCategoryModel.findById(
+        board.category.categoryId,
+      );
+      if (!oldCategory) {
+        throw new HttpError("Error updating board", 500);
+      }
+
+      const idx = oldCategory.boards.findIndex(
         (boardObj) => boardObj.id === boardId,
       );
 
-      oldCategory?.boards.splice(idx!, 1);
-      await oldCategory?.save({ session: sess, validateModifiedOnly: true });
+      oldCategory.boards.splice(idx, 1);
+      await oldCategory.save({ session: sess, validateModifiedOnly: true });
 
       // Add to new category
       const newCategory = await BoardCategoryModel.findById(categoryId);
-      newCategory?.boards.push(board._id);
-      await newCategory?.save({ session: sess, validateModifiedOnly: true });
+      if (!newCategory) {
+        throw new HttpError("Could not find category", 404);
+      }
 
-      board.category = { id: newCategory!._id, topic: newCategory!.topic };
+      newCategory.boards.push(board._id);
+      await newCategory.save({ session: sess, validateModifiedOnly: true });
+
+      board.category._id = newCategory._id;
+      board.category.topic = newCategory.topic;
     }
     await board.save({ session: sess, validateModifiedOnly: true });
 
@@ -109,9 +117,9 @@ const update = async (
 /**
  * Create a new board
  *
- * @param topic The name of the new board
- * @param description A brief description of the board's topic
- * @param categoryId The id of the category the board belongs to
+ * @param {string} topic - The name of the new board
+ * @param {string} description - A brief description of the board's topic
+ * @param {string} categoryId - The id of the category the board belongs to
  *
  * @returns An object containing board information
  */
@@ -148,7 +156,7 @@ const create = async (
     topic,
     description,
     category: {
-      id: categoryId,
+      categoryId: categoryId,
       topic: category.topic,
     },
   });
@@ -158,9 +166,8 @@ const create = async (
     sess.startTransaction();
 
     await board.save({ session: sess });
-    const boardId = (board.toObject({ getters: true }) as any).id;
 
-    category.boards.push(boardId);
+    category.boards.push(board._id);
     await category.save({ session: sess, validateModifiedOnly: true });
 
     await sess.commitTransaction();
@@ -174,7 +181,7 @@ const create = async (
 /**
  * Delete a board from the database
  *
- * @param {string} id The id of the board to delete
+ * @param {string} id - The id of the board to delete
  */
 const del = async (id: string) => {
   let board;
@@ -195,12 +202,17 @@ const del = async (id: string) => {
 
     // Remove the board from its category
     const boardId = board.id;
-    const oldCategory = await BoardCategoryModel.findById(board.category.id);
-    const idx = oldCategory?.boards.findIndex(
+    const oldCategory = await BoardCategoryModel.findById(
+      board.category.categoryId,
+    );
+    if (!oldCategory) {
+      throw new HttpError("Error deleting board", 500);
+    }
+    const idx = oldCategory.boards.findIndex(
       (boardObj) => boardObj.id === boardId,
     );
-    oldCategory?.boards.splice(idx!, 1);
-    await oldCategory?.save({ session: sess, validateModifiedOnly: true });
+    oldCategory.boards.splice(idx, 1);
+    await oldCategory.save({ session: sess, validateModifiedOnly: true });
 
     // Delete the board
     await BoardModel.findByIdAndDelete(boardId);
